requirements.txt
Django==3.2.8
gunicorn==20.0.4
django_environ==0.7.0
dj_database_url==0.5.0
requests==2.26.0
Faker==9.3.1
django_heroku==0.3.1
djangorestframework==3.12.4
coreapi==2.3.3
coreschema==0.0.4
urllib3==1.26.7
whitenoise==5.3.0
requests==2.26.0
factory-boy==3.1.0
PyYAML==5.3.1
isort==5.9.3